frame: add sequence number checks for connection IDs
Connection ID sequence numbers are stored as u32, which is sufficient, but
extra validation is needed to prevent overflow and invalid values when
processing retire_conn_id and new_conn_id frames or modifying a conn_id.

Also switch to u64 in quic_outq_transmit_new_conn_id() to stay consistent
with quic_outq_transmit_retire_conn_id().

+++ modules/net/quic/outqueue.c
@@ -1153,7 +1153,7 @@ int quic_outq_transmit_frame(struct sock *sk, u8 type, void *data, u8 path, u8 c
 int quic_outq_transmit_new_conn_id(struct sock *sk, u64 prior, u8 path, u8 cork)
 {
 	struct quic_conn_id_set *id_set = quic_source(sk);
-	u32 max, seqno;
+	u64 max, seqno;
 
 	/* Compute the maximum sequence number to send. */
 	max = id_set->max_count + prior;
